requirements.txt
# LLM-RL Framework Dependencies

# Core dependencies
gymnasium>=0.29.0
minigrid>=2.3.0
numpy>=1.24.0
pandas>=2.0.0

# LLM interface
ollama>=0.1.0
requests>=2.31.0

# RL baselines
stable-baselines3>=2.1.0
torch>=2.0.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and monitoring
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
